+++ CHANGELOG.md
@@ -25,6 +25,7 @@ The document follows the conventions described in [“Keep a CHANGELOG”](http:
 - added `SVG.Style` which can be created with `style()` or `fontface()` (#517)
 - added `EventTarget` which is a baseclass to get event abilities (#641)
 - added `Dom` which is a baseclass to get dom abilities
+- added `round()` which lets you round attribues from a node
 
 ### Removed
 - removed `SVG.Array.split()` function